backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
google-auth==2.35.0
google-api-python-client==2.149.0
pytest==8.3.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
gunicorn==23.0.0
httpx==0.27.2
pytest-mock==3.14.0
pytest-cov==5.0.0
google-cloud-storage==2.18.2
python-multipart==0.0.10